[YS-430] feat: CSR환경과 SSR환경에서 사용하는 fetchClient 분리
